Move Either tests into the Module tests.
